+++ NEWS.md
@@ -1,6 +1,7 @@
 ## Version 0.9.2
 
 * Added a `label = c("name", "formula")` argument to `LRstats()` to provide for labeling models by their model formulas in the output using `get_models()`.
+* Handle list (...) of models with formula labels more flexibly in `LRstats()`
 
 
 ## Version 0.9.1 (2026-02-08)